feat: ajout de la gestion CRUD de tous les UC + gestion de la bdd et ajout d'enregistrement par défaut.

+++ Lib_Entities/Entities/Espace.cs
@@ -26,7 +26,7 @@ namespace Lib_Entities.Entities
         /// <summary>
         /// Surface totale de l'espace en unités mètres carrées .
         /// </summary>
-        public double Superficie { get; set; }
+        public int Superficie { get; set; }
 
         /// <summary>
         /// Capacité maximale d'accueil (nombre maximal de personnes).
